executor/requirements.txt
aiokafka>=0.11
python-dotenv>=1.0
metaapi-cloud-sdk>=27.0
